Issue-3 Rework some plugin plumbing following IDWR feedback
1) Return CannotParse() if an exception occurs before one line is encountered.

This will allow the plugin to gracefully skip text files that are not a match for the expected CSV (like JSON, or other CSVs)

2) Added a configurable Survey.LocationAliases dictionary

This allows a remapping of location Identifiers in the CSV stream to different AQTS location identifiers.

3) Change some column mapping

Don't import the logger readings or battery readings.
Add the Measured discharge when it exists.