PF-349 - More fixes after testing
1) Reworked time parsing, to allow the visit time to be inferred from an activity if no explicit visit time is specified

2) Removed the concept of a 'default' property

It just didn't make sense, in documentation or in TOML syntax.

3) Improved preface definition options

Added PrefaceMustContain and PrefaceMustMatchRegex properties, to allow for quick rejection of files which should not be parsed.

4) Improved header area parsing

Added HeaderRowCount property, to allow for a multi-line header (common for Excel files).

5) Improved the column shorthand syntax, to allow aliases.

Now most every variation of FixedValue, ColumnHeader, ColumnIndex, or PrefaceRegex can be set with a string.

And optional aliases can be set at the end of the string.