PF-349 - More plugin fixes
1) Configurations with Priority <= 0 are considered disabled and are not validated or processed

2) Added Configuration.NoDataRowsExpected property

Set this to true when all the field visit data comes from the Preface (say, if the text file is generated from a form)

3) The default TimestampDefinition.TimestampType is DateTimeOnly

The default timestamp now assumes times are local to the location, which seems to be most common.

4) Support multiple timestamp Formats

The TimestampDefinition.Formats property is now an array, and supports multiple format specifications if needed.

5) The default timestamp format is the .NET "try all the culture specific formats"

The TimestampDefinition.Formats property is empty by default.

When empty, use the .NET DateTimeOffset.TryParse() method which attempts all the culture-specific date time formats.

This is often good enough for most parsing scenarios.