PF-1438 - Some quality-of-life improvements to support EPCAMR's FV data ingest
1) SkipVisitIfNoActivities = true

This configuration flag can be set to skip a visit that may have a dates and comment, but no other activities (because all the readings are empty)

2) Allow an alias from an empty string to force a short circuit

An alias from an empty string to anything was never useful in the past.

Now we treat an empty-string alias as an "else clause", forcing anything that doesn't match to that value.
This will help quite a bit of data wrangling in general

3) Create an empty reading if the value column was not empty, but was aliased to an empty string.

This will allow a Reading.Value = '@Column{Alias}' to work with an alias of non-detect values.